added basic enemyFish movement, made fish.move more generic, modified initial fish coordinates
Added object enemyFish, modified fish.move to make it look at
conditions specific to each fish (moveLeft, right, up, and down) that
are now determined each tick in checkMove. To fix: if enemyFish is at
the same x and/or y coordinate as fish, due to the checks used for
enemyFish movement and general fish physics, enemyFish “jitters” rather
than staying still. Also modified the fish starting coordinates so that
it starts at a random y value, and either the left or right side of the
screen. To fix: if starting at the right, the fish starts off
immediately off-screen; the position needs to be offset by the fish
image’s width.