Big changes to fish movement
Reverted fish.move to be player-controlled-fish specific, removed
checkMove from renderCanvas, started adding smarter movement for
enemyFish (with the Fish.hunt() function):
enemyFish finds a straight, diagonal line towards fish and travels
along it, using trig to convert the calculated vector magnitude into x
and y components. Also added a deceleration aspect: when the enemyFish
is in a place where it can make its x or y coordinate be equal to that
of fish by decelerating naturally and opposing its own momentum, it
will do so. This (mostly) fixes the issue of enemyFish “jittering” when
near the same x and/or y coordinate of fish. There is still an issue
with this new function; when moving in x or y negative directions, its
speed can become faster than its maxSpeed. This function still needs to
be cleaned up/polished.
Additionally, updated cleanFish’s splice() command, adding a second
argument of 1 (as it defaults to 0, causing an issue with spawning new
dumb fish).